BuergerRWForm replaces Update and ReadForm
